Move utils functions to init
